Rename `Wings::ResourceFactory` to `Wings::ModelTransformer`
The `ResourceFactory` name collides with a Valkyrie concept, creating potential
for confusion. This rename is slightly more specific about what this class is
for: transforming one kind of model into another.

The old name is retained to ease migration for dependant apps.
