+++ lib/wings.rb
@@ -7,6 +7,7 @@ module Wings
 end
 
 require 'valkyrie'
+require 'wings/model_transformer'
 require 'wings/resource_factory'
 require 'wings/valkyrizable'
 require 'wings/valkyrie_monkey_patch'
