+++ spec/wings/model_transformer_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 require 'spec_helper'
-require 'wings/resource_factory'
+require 'wings/model_transformer'
 require 'wings/valkyrie_monkey_patch'
 
-RSpec.describe Wings::ResourceFactory do
+RSpec.describe Wings::ModelTransformer do
   subject(:factory) { described_class.new(pcdm_object: pcdm_object) }
   let(:pcdm_object) { work }
   let(:adapter)     { Valkyrie::MetadataAdapter.find(:memory) }
